Break machine and network into seperate files
