+++ services/shared/network-configuration/index.ts
@@ -1,0 +1,4 @@
+import { Machine } from './src/machine'
+import { Network } from './src/network'
+
+export { Machine, Network }
