+++ services/shared/network-configuration/package.json
@@ -1,8 +1,8 @@
 {
   "name": "shared-library",
   "version": "1.0.0",
-  "main": "dist/network-configuration.js",
-  "types": "dist/network-configuration.d.ts",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "license": "MIT",
   "private": true,
   "scripts": {
